Add an update script and a placeholder for the old am to allow update
The old access method was called embedding. Now it is called hnsw.
In the databases where the old extension is installed we cannot
drop the extension because for some reason postgres wants to
access the c definition of the access method before dropping

This commit adds an update script and adds a placeholder in
the shared library for the old access method so
ALTERN EXTENSION lanterndb UPDATE; can bring in the new version