build.c: use inserts to populate index on creation
Add a feature-toggled way to have build.c populate hnsw
index using insert logic.

The reason for this is two folds:
- Unify build and insertion logic
- Upcoming migration away from blockmap, and the
  current build logic's incompatibility with it

Note that the change here for build.c only applies to the case
when an index is created on a table that already has rows.

In particular, externally created index getting uploaded to
postgres will not cause the new logic to get exercised.

Feature is disabled by default for now as we haven't
benchmarked this new approach, and think it'll
impact the performance of index creation.