Update flaky regression tests so ci/cd passes
None of regressions in these tests are substantial.
We can avoid these kind of things in the future by making
tests more robust.

There are the following kind of changes

1) row order in select changed

returned rows are sorted by distance and the changed order
has the same distance values. Multiple rows with same distance
value are present in test output and ties are broken arbitrarily
The arbitrary tie-break causes a regression

2) cost estimate output is different

Perhaps postgres estimates a slightly differnet number of rows
in the table which causes this regression. Not sure how we can harden it