Add blocks around switch cases to adhere the C standard
Per the standard:
https://port70.net/~nsz/c/c11/n1570.html#6.8.1

Labels must be followed by a statement.
Switch 'case's are just labels.
So, we cannot have a variable declaration right after a 'case'.
We *can* have a code block however, which is what this PR does.

It seems the newer compilers handle the case when a declaration comes
after a switch 'case'. But my compiler (gcc 10.2) complained about it