Make check_symbols not accept dynamic c++ symbols
Turns out our static usearch library still depended on the cpp runtime
library libstdc++ dynamically.
This usually did not cause issues since postgres
commonly links libstdc++ dynamically so c++ symbols in our .so would be
resolved dynamically, when loaded.

This did cause issues for when postgres was not linked with libstdc++. I
do not know when exactly this happens but it seems this at least happens
when building postgres from source under certain circumstances

This commit makes sure that we do not use libstdc++ symbols from
postgres dynamic library dependency and actually enforce that all c++
symbols we require are in the sahred object we build