Add update routine that handles index format changes
So far update scripts only made changes in the SQL layer
This works so long as the data structures in the old Lantern and
new lantern versions are compatible.

This is not the case when we change storage format, or header page,
for example.

To mitigate this,we can either:
1. Write binary level update routines that patch necessary DB and lanter
 index pages
2. Recreate all binary data structures by rerunning equivalent SQL-level
 statements

This commit implements the second approach. We can stil have one-off
cases of the first. But, whenever something in the binary formats of
lantern chagnes and a specific migration script (1) does not exist,
the 2nd approach must appropriately be integrated, as demonstrated in
this commit