Do not call hooks during update and creation
Now that we preload lantern shared library, our hooks see more
statements than before.
With this we ignore create/update statements in hooks because
1) they are irrelevant
2) our hooks assume DB state that is not satisfied when these
statements are running

Warn when lantern is preloaded.

Note: These checks were added when considering a move to preloaded
lantern. We ultimately abandoned the move, so, the code is only relevant
in self-hosting if one decides to preload lantern.