Increase m and therefore index size in hnsw_create_* to avoid size randomness
Newer storage made the index slightly smaller which
put old m=4 sift 1k index be near the border of a filled
index page. As a result, depending on level distribution
in a run, the regression test would have a differently sized
index and result in a failure.

Here we increase m slighly, to make it very unlikely a test will
fail as a result of level randomness