Fix snapshot leak bug
In testing we never had other libraries that modified the ExecutorStart
hook.
As a result we did not catch the bug that got triggered only when were
not the only one modifying the hook.
This commit reproduces the issue in testing and fixes it.

The problem was that we were calling the original executor start
hook even when there was another hook present, loaded by another
extension.

But the right approach is to only call that other hook and assumes
that hook itself will call the standard hook down the chain.

See pg_stat_statements implementation of the relevant section here:
Lines 991-994:
https://doxygen.postgresql.org/pg__stat__statements_8c.html#af004be8a1c6880821a28272be9a2b205