Add `external` param to index creation. Add `lantern_hnsw.external_index_host` GUC. Add `lantern_hnsw.external_index_port` GUC.
When the `external` parameter will be passed we connect to remote lantern daemon  via tcp socket using the `external_index_host` and `external_index_port` GUC variables.
First we send the index header to socket and codebook if needed. After getting ack response we start streaming tuples via socket to daemon which will create multithreaded usearch index. Then it will send back the `num_tuples_added` and `index_size` (which is the file size for index) and start streaming index file.
We will collect index file into `result_buf` and use that to write the index into disk.
For now we assume that the data sent will be in little endian byte ordering.